Ensured we use longs everywhere (for no real justifiable reason)

+++ src/JSONAbstractDescription.cpp
@@ -52,8 +52,8 @@ std::shared_ptr<AbstractDualGraph> JSONAbstractDescription::toAbstractDualGraph(
   AbstractDualGraph adg(d_zones.size());
   std::shared_ptr<AbstractDualGraph> graph = std::make_shared<AbstractDualGraph>(adg);
 
-  for(int i=0; i < d_zones.size(); i++) {
-    for(int j=i; j < d_zones.size(); j++) {
+  for(unsigned long i=0; i < d_zones.size(); i++) {
+    for(unsigned long j=i; j < d_zones.size(); j++) {
       if(1 == hammingDist(d_zones[i], d_zones[j])) {
         add_edge(i, j, *graph);
       }
